[DR-2137] Check billing profile access when creating snapshots (#1010)
* Improve snapshot creation usability

Disable the snapshot create button if a
user does not have access to the billing
profile. This does not yet handle the
scenario where admins can have view access
but not linking access.

* Conditionally show snapshot sharing panel

If a user cannot create a snapshot, do not
render the panel to share it.

* Update create snapshot button style

Make the button blue when active. It
will remain gray when disabled.

* Fix linting errors

* Fix more eslint errors

* Fix cypress tests

* Fix linting

* Fix billing profile check

* Refactor fetching billing profile

* Fix linting

* Code review comments