Fix bug in authPut function
I think axios.put was not finding the headers and specifically the auth token where it expected because params were included with the headers.
It interpreted the entire blob as its request payload and responded with an 'invalid_request' bearer error (no bearer token found in the request).
axios.put has the following method alias: axios.put(url[, data[, config]])

So we should pass in the request URL, the params (data), and the header blob.
Verified that I could update a snapshot's DUOS ID locally with this change.