EditableField loading behavior while awaiting response
- Save and cancel buttons are disabled
- Save button displays loading spinner
- Field cannot be edited

Also added handling specific to the case where we unset a snapshot's DUOS ID: the input from the text box is an empty string, but the returned object on successful unset has a null DUOS ID.

Additional work is needed to handle the case where the response is unsuccessful: right now, the system still thinks that the save is pending, so the buttons and edit capability remain disabled.  The only way out is to refresh the page.