[DR-2928] Collapse all resource role accordions by default (#1291)
Previously, dataset and snapshot stewards were expanded by default when clicking into the resource summary's roles and memberships tab.
We've made this change because sometimes the stewards list can be long enough to annoy the user by keeping it open by default.
We considered also including a member count in the collapsed accordion headers -- i.e. 'Stewards (34)' -- but decided against it because such a count would only reflect direct members and could confuse our users.  Specifically, if a Firecloud group is a direct member of a resource's policy group, we won't be able to tell how many members it has.