[DR-2969] Display DUOS group's last synced time
If a snapshot is linked to a DUOS dataset, we now display the last synced time of its Firecloud group in the Snapshot Summary panel.
If the Firecloud group has never been synced (shouldn't happen as we sync on group creation), we indicate as much.
Expanded unit tests for these cases.