[DCJ-572] Hide Activity tab
For non-Admin users, it's currently timing out and driving up load balancer latency stats, which pages our On Call engineers.

This tab is clicked on only sporadically: this PR hides it so that users don't stumble upon it while we debug the underlying slowness.

Users can if needed still navigate to the /activity URL to view all jobs, but only TDR Admins / debuggers are likely to actually do that.